app/src/main/java/io/barnabycolby/sqrlclient/activities/NoIdentityActivity.java
package io.barnabycolby.sqrlclient.activities;

import android.content.Intent;
import android.os.Bundle;
import android.support.v7.app.AppCompatActivity;
import android.view.View;

import io.barnabycolby.sqrlclient.activities.CreateNewIdentityActivity;
import io.barnabycolby.sqrlclient.App;
import io.barnabycolby.sqrlclient.R;
import io.barnabycolby.sqrlclient.sqrl.SQRLIdentityManager;

/**
 * Activity displayed when no SQRL identities are associated with the application, for example, the first time the application is launched.
 */
public class NoIdentityActivity extends AppCompatActivity {
    @Override
    protected void onCreate(Bundle savedInstanceState) {
        // Standard Android stuff
        super.onCreate(savedInstanceState);
        setContentView(R.layout.activity_no_identity);
    }

    @Override
    protected void onResume() {
        super.onResume();

        // Check whether there really are no identities
        SQRLIdentityManager identityManager = App.getSQRLIdentityManager();        
        if (identityManager.containsIdentities()) {
            Intent intent = new Intent(this, MainActivity.class);
            startActivity(intent);
        }
    }

    /**
     * Called when the "Create New Identity" button is pressed.
     */
    public void createNewIdentity(View view) {
        Intent intent = new Intent(this, CreateNewIdentityActivity.class);
        startActivity(intent);
    }
}
